web/requirements.txt
Django==3.2.10
pip==21.3.1
wheel==0.37.0
pytz==2021.3
sqlparse==0.4.2
psycopg2==2.9.2
asgiref==3.4.1
setuptools==58.5.3
djangorestframework==3.12.4